requirements.txt
selenium
webdriver-manager
python-telegram-bot>=20.0
apscheduler
python-dotenv
aiosqlite
